+++ cmd/swisstronikd/cmd/config_test.go
@@ -1,4 +1,4 @@
-package client
+package cmd
 
 import (
 	"os"
