+++ cmd/swisstronikd/cmd/keys/add.go
