+++ cmd/swisstronikd/cmd/keys/utils.go
